pyproject.toml
[tool.poetry]
name = "muscatplus-reconciliation"
version = "0.1.0"
description = ""
authors = ["Andrew Hankinson <andrew.hankinson@gmail.com>"]
readme = "README.md"
packages = [{include = "muscatplus_reconciliation"}]

[tool.poetry.dependencies]
python = "^3.11"
sanic = "^23.3.0"
small-asc = {git = "https://github.com/rism-digital/small-asc.git", rev = "main"}
ypres = {git = "https://github.com/rism-digital/ypres.git", rev = "main"}
orjson = "^3.8.11"


[tool.poetry.group.dev.dependencies]
ipython = "^8.13.1"
sanic-testing = "^23.3.0"
pytest-asyncio = "^0.21.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
